refactor(test): Move ExtendedStringLiteralTestClass to testclasses (#2153)
